pyproject.toml
[project]
name = "nolas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "imapclient>=3.0.1",
    "requests>=2.32.4",
    "asyncpg>=0.29.0",
    "aioimaplib>=2.0.1",
    "aiohttp>=3.9.5",
    "pydantic-settings>=2.10.1",
    "python-json-logger>=3.3.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.0",
    "greenlet>=3.2.3",
    "dependency-injector>=4.48.1",
    "fastapi>=0.104.0",
    "fastapi-async-sqlalchemy>=0.6.0",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.4.0",
    "mypy>=1.10.0",
    "ruff>=0.4.0",
]

[tool.mypy]
python_version = "3.13"
plugins = ["pydantic.mypy"]
strict = true
ignore_missing_imports = true
disallow_untyped_decorators = false
disallow_untyped_calls = false
disallow_subclassing_any = false
implicit_reexport = false
warn_unused_ignores = false
exclude = "(.venv)/$"


[tool.black]
line-length = 120

[tool.ruff]
line-length = 120
